requirements.txt
cycler==0.10.0
fpdf==1.7.2
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
Pillow==6.2.1
py3dns==3.2.1
pyparsing==2.4.4
python-dateutil==2.8.1
six==1.13.0
validate-email==1.3
